refactor(parser): extract the reparse module
`parse_incremental_suffix[_with_refdefs]` and `IncrementalParseResult`
retire: nothing in production used them after the reparse moved into
salsa, and they silently full-parsed, which is exactly the ambiguity the
refusal-first shape removes.

The machinery they wrapped -- the guard cascade, the window finders, the
splice, the error merge -- moves out of the 1500-line `parser.rs` into
`parser/reparse.rs` beside `reparse` itself. Pure code motion; the
roadmap scheduled it for the token tier, but S4 was already moving this
code.

The "splice if you can, else full-parse" policy now lives with the
callers that want it: a shared `tests/common` module for the four
incremental suites, a `#[cfg(test)]` shim for the in-crate tests, and
`try_reparse` in the bench, where the decline path is a measurement.

Docs record the one-tree architecture and the new
`PANACHE_INCREMENTAL_PARSING` override; TODO.md checks off phase 4.
